feat: preserve last frame when stream ends
The `video` element is tightly bound to the camera stream. If the
stream ends the `video` is stopped too. By using the `video` as
a mediator and painting it's frames to a visible canvas, the last
frame of a camera stream can be preserved even when the stream
ends.

Close #24